# Improve series deduplication for `/active_series` endpoint