--- a/pkg/distributor/distributor_test.go
+++ b/pkg/distributor/distributor_test.go
@@ -2178,6 +2178,8 @@ func TestDistributor_MetricsForLabelMatchers(t *testing.T) {
 func TestDistributor_ActiveSeries(t *testing.T) {
 	const numIngesters = 5
 
+	collision1, collision2 := labelsWithHashCollision()
+
 	pushedData := []struct {
 		lbls      labels.Labels
 		value     float64
@@ -2186,6 +2188,8 @@ func TestDistributor_ActiveSeries(t *testing.T) {
 		{labels.FromStrings(labels.MetricName, "test_1", "team", "a"), 1, 100000},
 		{labels.FromStrings(labels.MetricName, "test_1", "team", "b"), 1, 110000},
 		{labels.FromStrings(labels.MetricName, "test_2"), 2, 200000},
+		{collision1, 3, 300000},
+		{collision2, 4, 300000},
 	}
 	tests := map[string]struct {
 		shuffleShardSize            int
@@ -2208,6 +2212,11 @@ func TestDistributor_ActiveSeries(t *testing.T) {
 			requestMatchers:             []*labels.Matcher{labels.MustNewMatcher(labels.MatchEqual, model.MetricNameLabel, "test_2")},
 			expectedSeries:              []labels.Labels{pushedData[2].lbls},
 			expectedNumQueriedIngesters: 3,
+		},
+		"should return all matching series even if their hash collides": {
+			requestMatchers:             []*labels.Matcher{labels.MustNewMatcher(labels.MatchEqual, model.MetricNameLabel, "metric")},
+			expectedSeries:              []labels.Labels{collision1, collision2},
+			expectedNumQueriedIngesters: numIngesters,
 		},
 	}
 
@@ -2653,6 +2662,25 @@ func TestDistributor_LabelNamesAndValues_ExpectedAllPossibleLabelNamesAndValuesT
 	require.NoError(t, err)
 	require.Len(t, response.Items, 1)
 	require.Equal(t, 10000, len(response.Items[0].Values))
+}
+
+// copied from pkg/ingester/activeseries/active_series_test.go
+func labelsWithHashCollision() (labels.Labels, labels.Labels) {
+	// These two series have the same XXHash; thanks to https://github.com/pstibrany/labels_hash_collisions
+	ls1 := labels.FromStrings("__name__", "metric", "lbl1", "value", "lbl2", "l6CQ5y")
+	ls2 := labels.FromStrings("__name__", "metric", "lbl1", "value", "lbl2", "v7uDlF")
+
+	if ls1.Hash() != ls2.Hash() {
+		// These ones are the same when using -tags stringlabels
+		ls1 = labels.FromStrings("__name__", "metric", "lbl", "HFnEaGl")
+		ls2 = labels.FromStrings("__name__", "metric", "lbl", "RqcXatm")
+	}
+
+	if ls1.Hash() != ls2.Hash() {
+		panic("This code needs to be updated: find new labels with colliding hash values.")
+	}
+
+	return ls1, ls2
 }
 
 func prepareWithZoneAwarenessAndZoneDelay(t *testing.T, fixtures []series) (context.Context, []*Distributor) {