<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does
This adds a benchmark for `distributor.ActiveSeries` and tries to improve efficiency of the code by doing less work to deduplicate series streaming in from ingesters.

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/distributor
                            │    main     │               branch               │
                            │   sec/op    │   sec/op     vs base               │
Distributor_ActiveSeries-10   17.34m ± 2%   11.84m ± 1%  -31.72% (p=0.002 n=6)

                            │     main      │               branch                │
                            │     B/op      │     B/op      vs base               │
Distributor_ActiveSeries-10   19.020Mi ± 0%   4.647Mi ± 0%  -75.57% (p=0.002 n=6)

                            │     main     │               branch               │
                            │  allocs/op   │  allocs/op   vs base               │
Distributor_ActiveSeries-10   135.32k ± 0%   85.36k ± 0%  -36.92% (p=0.002 n=6)
```

#### Which issue(s) this PR fixes or relates to

n/a

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.