@colega 2023-11-24
When running with `stringlabels` (which is how we run Mimir now) most of the time is spent doing `FromLabelAdaptersOverwriteLabels`:

![image](https://github.com/grafana/mimir/assets/1511481/4c6d0b39-ca30-42f4-b69f-cbaebea3e505)

So, to further optimize this, I'd copy the non-stringlabels implementations of `Labels.Hash()` to `LabelAdapters.Hash()` and remove the useless conversion to stringlabels.