@colega 2023-11-24
> The hash isnt guaranteed to be unique, right? So technically this could miss some series

Yes. There are known collisions.