@flxbk 2023-11-27
> The hash isnt guaranteed to be unique, right? So technically this could miss some series

Apart from using a larger hash or the label string itself as the key for deduplication, does anything come to mind for how we could circumvent this problem? I'd assume that for the cardinality use case, the error introduced by this would be insignificant, wdyt?