@flxbk 2023-11-28
I added a metric to count hash collisions, which we could divide by the number of series returned from this API overall to get a sense of how big the impact is. If the problem seems large enough to worry about we can think about collision resolution, wdyt?