@colega 2023-11-28
@dimitarvdimitrov 
> we could end up implementing a map where the type of m.series is map[uint64][]labels.Labels. That would create a lot of extra garbage though for all the slices.

We can allocate a slice only when we see collisions:

```go
type entry struct {
    first labels.Labels
    collisions []labels.Labels
}

series := map[uint64]entry{}

for ls := range foobar {
    hash := ls.Hash()
    if e, ok := series[hash]; ok {
        e.collisions = append(e.collisions, ls)
        series[hash] = e
    } else {
        series[hash] = entry{first: ls}
    }
}
```